% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.R
\name{citations}
\alias{citations}
\title{function that returns citations for a list of packages}
\usage{
citations(packages = NULL, type = c("text", "bibtex", "raw", "html",
  "markdown", "list"), file = NULL)
}
\arguments{
\item{packages}{if NULL the packages_in_use() function is used genarate a
list of all packages currently in use and work with that; if a character
vector of package names is supplied all those package will be cited}

\item{type}{the return type, only first element will be used: (raw) the raw
values returned by citation() are returned as a list; (bibtex) a character
vector of BibTeX entries is returned with each entry coresponding to one
element; (text) a character vector of un-marked-up text is returned with
each element coresponding to one citation; (markdown) a character vector of
preformatted markdown text; (html) a character vector of preformatted html
text}
}
\description{
function that returns citations for a list of packages
}

