% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_bibtype}
\alias{get_bibtype}
\title{function to extract bibtype attribute from citation object}
\usage{
get_bibtype(x)
}
\arguments{
\item{x}{citation object to get textVersion from}
}
\description{
function to extract bibtype attribute from citation object
}

