% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_text_version}
\alias{get_text_version}
\title{function to extract textVersion attribute from citation object}
\usage{
get_text_version(x)
}
\arguments{
\item{x}{citation object to get textVersion from}
}
\description{
function to extract textVersion attribute from citation object
}

