% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{packages_in_use}
\alias{packages_in_use}
\title{function returning character vector of packages currently in use}
\usage{
packages_in_use(base = TRUE, loadedonly = TRUE, attached = TRUE)
}
\arguments{
\item{base}{should names of base packages be returned?}

\item{loadedonly}{should loaded but not attached packeges be returned?}

\item{attached}{should names of attached, non-bse packages be returned?}
}
\description{
function returning character vector of packages currently in use
}

