% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{set_class}
\alias{set_class}
\title{function returning x with modified class}
\usage{
set_class(x, y)
}
\arguments{
\item{x}{object for which class should be changed}

\item{y}{character vector of class names to which to switch}
}
\description{
function returning x with modified class
}

