% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{to_bibtex_string}
\alias{to_bibtex_string}
\title{function transforming citation object to bibtex string}
\usage{
to_bibtex_string(x)
}
\arguments{
\item{x}{citation object for which to produce Bibtex string}
}
\description{
function transforming citation object to bibtex string
}

